# packages/contracts/openapi/openapi.json
{
  "components": {
    "schemas": {
...
      "Body_upload_skill_version_endpoint_admin_api_v1_skills__skill_id__versions_post": {
        "properties": {
          "package": {
            "contentMediaType": "application/octet-stream",
            "title": "Package",
            "type": "string"
          },
          "version": {
            "title": "Version",
            "type": "string"
          }
        },
        "required": [
          "version",
          "package"
        ],
        "title": "Body_upload_skill_version_endpoint_admin_api_v1_skills__skill_id__versions_post",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "SkillCreate": {
        "additionalProperties": false,
        "properties": {
          "description": {
            "maxLength": 20000,
            "minLength": 1,
            "title": "Description",
            "type": "string"
          },
          "display_name": {
            "maxLength": 100,
            "minLength": 1,
            "title": "Display Name",
            "type": "string"
          },
          "key": {
            "maxLength": 100,
            "minLength": 1,
            "pattern": "^[a-z0-9][a-z0-9_-]*$",
            "title": "Key",
            "type": "string"
          }
        },
        "required": [
          "key",
          "display_name",
          "description"
        ],
        "title": "SkillCreate",
        "type": "object"
      },
      "SkillResponse": {
        "additionalProperties": false,
        "properties": {
          "catalog_revision": {
            "title": "Catalog Revision",
            "type": "integer"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "description": {
            "maxLength": 20000,
            "minLength": 1,
            "title": "Description",
            "type": "string"
          },
          "display_name": {
            "maxLength": 100,
            "minLength": 1,
            "title": "Display Name",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "key": {
            "maxLength": 100,
            "minLength": 1,
            "pattern": "^[a-z0-9][a-z0-9_-]*$",
            "title": "Key",
            "type": "string"
          },
          "status": {
            "enum": [
              "draft",
              "validated",
              "published"
            ],
            "title": "Status",
            "type": "string"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          },
          "version": {
            "title": "Version",
            "type": "integer"
          }
        },
        "required": [
          "key",
          "display_name",
          "description",
          "id",
          "catalog_revision",
          "status",
          "created_at",
          "updated_at",
          "version"
        ],
        "title": "SkillResponse",
        "type": "object"
      },
      "SkillUpdate": {
        "additionalProperties": false,
        "properties": {
          "description": {
            "anyOf": [
              {
                "maxLength": 20000,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "display_name": {
            "anyOf": [
              {
                "maxLength": 100,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Display Name"
          }
        },
        "title": "SkillUpdate",
        "type": "object"
      },
      "SkillVersionResponse": {
        "additionalProperties": false,
        "properties": {
          "content_sha256": {
            "title": "Content Sha256",
            "type": "string"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "manifest_json": {
            "additionalProperties": true,
            "title": "Manifest Json",
            "type": "object"
          },
          "object_key": {
            "title": "Object Key",
            "type": "string"
          },
          "object_sha256": {
            "title": "Object Sha256",
            "type": "string"
          },
          "object_size": {
            "title": "Object Size",
            "type": "integer"
          },
          "published_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Published At"
          },
          "signature": {
            "title": "Signature",
            "type": "string"
          },
          "signing_key_id": {
            "title": "Signing Key Id",
            "type": "string"
          },
          "skill_id": {
            "format": "uuid",
            "title": "Skill Id",
            "type": "string"
          },
          "version": {
            "title": "Version",
            "type": "string"
          }
        },
        "required": [
          "id",
          "skill_id",
          "version",
          "manifest_json",
          "object_key",
          "object_size",
          "object_sha256",
          "signature",
          "signing_key_id",
          "content_sha256",
          "published_at",
          "created_at"
        ],
        "title": "SkillVersionResponse",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "title": "iBreeze Backend API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/admin/api/v1/skills": {
      "get": {
        "operationId": "list_skills_endpoint_admin_api_v1_skills_get",
        "parameters": [
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "maximum": 200,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response List Skills Endpoint Admin Api V1 Skills Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "List Skills Endpoint",
        "tags": [
          "skills"
        ]
      },
      "post": {
        "operationId": "create_skill_endpoint_admin_api_v1_skills_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SkillCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SkillResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Create Skill Endpoint",
        "tags": [
          "skills"
        ]
      }
    },
    "/admin/api/v1/skills/{skill_id}": {
      "delete": {
        "operationId": "delete_skill_endpoint_admin_api_v1_skills__skill_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "skill_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Skill Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "If-Match",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "If-Match"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Delete Skill Endpoint",
        "tags": [
          "skills"
        ]
      },
      "get": {
        "operationId": "get_skill_endpoint_admin_api_v1_skills__skill_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "skill_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Skill Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SkillResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get Skill Endpoint",
        "tags": [
          "skills"
        ]
      },
      "patch": {
        "operationId": "update_skill_endpoint_admin_api_v1_skills__skill_id__patch",
        "parameters": [
          {
            "in": "path",
            "name": "skill_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Skill Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "If-Match",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "If-Match"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SkillUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SkillResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Update Skill Endpoint",
        "tags": [
          "skills"
        ]
      }
    },
    "/admin/api/v1/skills/{skill_id}/revisions": {
      "post": {
        "operationId": "clone_skill_endpoint_admin_api_v1_skills__skill_id__revisions_post",
        "parameters": [
          {
            "in": "path",
            "name": "skill_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Skill Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SkillResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Clone Skill Endpoint",
        "tags": [
          "skills"
        ]
      }
    },
    "/admin/api/v1/skills/{skill_id}/validate": {
      "post": {
        "operationId": "validate_skill_endpoint_admin_api_v1_skills__skill_id__validate_post",
        "parameters": [
          {
            "in": "path",
            "name": "skill_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Skill Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SkillResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Validate Skill Endpoint",
        "tags": [
          "skills"
        ]
      }
    },
    "/admin/api/v1/skills/{skill_id}/versions": {
      "get": {
        "operationId": "list_skill_versions_endpoint_admin_api_v1_skills__skill_id__versions_get",
        "parameters": [
          {
            "in": "path",
            "name": "skill_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Skill Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response List Skill Versions Endpoint Admin Api V1 Skills  Skill Id  Versions Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "List Skill Versions Endpoint",
        "tags": [
          "skills"
        ]
      },
      "post": {
        "operationId": "upload_skill_version_endpoint_admin_api_v1_skills__skill_id__versions_post",
        "parameters": [
          {
            "in": "path",
            "name": "skill_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Skill Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_upload_skill_version_endpoint_admin_api_v1_skills__skill_id__versions_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SkillVersionResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Upload Skill Version Endpoint",
        "tags": [
          "skills"
        ]
      }
    },
    "/admin/api/v1/skills/{skill_id}/versions/{version_id}": {
      "delete": {
        "operationId": "delete_skill_version_endpoint_admin_api_v1_skills__skill_id__versions__version_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "skill_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Skill Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "version_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Version Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Delete Skill Version Endpoint",
        "tags": [
          "skills"
        ]
      }
    },
...
    "/api/v1/catalog/skills/{skill_id}/versions/{version}/package": {
      "get": {
        "operationId": "download_skill_package_endpoint_api_v1_catalog_skills__skill_id__versions__version__package_get",
        "parameters": [
          {
            "in": "path",
            "name": "skill_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Skill Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "version",
            "required": true,
            "schema": {
              "title": "Version",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Download Skill Package Endpoint",
        "tags": [
          "skills"
        ]
      }
    },
...
  }
}